cambios en vistas para mejor organizacion
